fix: fix windows support
Sadly ppl lirary is not available for windows hence core won't be available for
windows users. But other features that do not depend on it can be still
supported. This checks if the pplpy (python wrapper around ppl) is available
and if not throws and error if user tries to use core. Also proprietly changes
the tests to skip them if core is not available.