git: drop testpypi and workflow
We can't reupload files with the same name to testpypi. This means that
to publish anything to testpypi we need to bump the version. Because of
this it looses purpose to publish to testpypi as we would need to bump
the version with every push to master, but we want to bump the version
just when publishing to pypi.