remove: problematic convex game generator
The library ``pyfmtools`` used for convex game generator is problematic. Not
only it was the only reason to have maximum python version 3.11, but it
not even always worked properly. Let's drop the convex game generator
for the time being to enable easier maintenance and possibility to use
newer version of the python.